server/pyproject.toml
[project]
name = "RAG-chat"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "uvicorn>=0.35.0",
    "fastapi>=0.116.1",
    "graypy>=2.1.0",
    "PyYAML>=6.0.2",
    "openai>=1.100.2",
    "PyPDF2>=3.0.1",
#    "PyMuPDF>=1.24.0",
    "jinja2>=3.1.0",
#    "qdrant_client>=1.15.1",
#    "sentence_transformers>=5.1.0"
]

[tool.ruff]
line-length = 160
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
